feat: front-load rehearsal manifest markdown for #1357
