+++ docs/phase1-candidate-rehearsal.md
@@ -86,6 +86,6 @@ npm run release:phase1:candidate-rehearsal -- \
   --target-surface h5
 ```
 
-Open `artifacts/release-readiness/phase1-candidate-rehearsal-local/SUMMARY.md` first. Its reviewer front-door section is the canonical packet-level reviewer entrypoint and directly calls out the current release evidence index, the release gate summary, the release gate summary Markdown companion, the release health summary, the release health summary Markdown companion, the CI trend summary, the CI trend summary Markdown companion, the release readiness snapshot, the runtime observability gate, the H5 candidate smoke, the reconnect soak summary, the WeChat candidate summary, the runtime observability bundle, the runtime observability evidence, candidate evidence audit, candidate freshness guard, candidate owner reminder, candidate freshness history, restaged release-readiness dashboard, the same-revision evidence bundle manifest, the Phase 1 release evidence drift gate, the Phase 1 exit audit, the Phase 1 exit-dossier freshness gate, the Phase 1 candidate dossier, the Phase 1 candidate dossier Markdown companion, the go/no-go packet, manual evidence owner ledger, Cocos primary journey evidence, the Cocos reconnect replay artifact, the Cocos main-journey replay gate, Cocos primary diagnostics, the candidate revision triage digest, the Cocos RC bundle, and the release PR summary, then records the release gate and release health outcomes for the candidate revision.
+Open `artifacts/release-readiness/phase1-candidate-rehearsal-local/SUMMARY.md` first. Its reviewer front-door section is the canonical packet-level reviewer entrypoint and now foregrounds the release candidate manifest Markdown artifact, followed by the manifest JSON companion, the current release evidence index, the release gate summary, the release gate summary Markdown companion, the release health summary, the release health summary Markdown companion, the CI trend summary, the CI trend summary Markdown companion, the release readiness snapshot, the runtime observability gate, the H5 candidate smoke, the reconnect soak summary, the WeChat candidate summary, the runtime observability bundle, the runtime observability evidence, candidate evidence audit, candidate freshness guard, candidate owner reminder, candidate freshness history, restaged release-readiness dashboard, the same-revision evidence bundle manifest, the Phase 1 release evidence drift gate, the Phase 1 exit audit, the Phase 1 exit-dossier freshness gate, the Phase 1 candidate dossier, the Phase 1 candidate dossier Markdown companion, the go/no-go packet, manual evidence owner ledger, Cocos primary journey evidence, the Cocos reconnect replay artifact, the Cocos main-journey replay gate, Cocos primary diagnostics, the candidate revision triage digest, the Cocos RC bundle, and the release PR summary, then records the release gate and release health outcomes for the candidate revision.
 
 For the standalone CI guard and explicit GitHub Actions inputs, see [`docs/phase1-release-evidence-drift-gate.md`](./phase1-release-evidence-drift-gate.md).
